back-end/requirements.txt
flask
numpy
sklearn
gunicorn
farasapy
pyarabic
requests
pandas
nltk

